requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
oauthlib==2.0.4
python-bitcoinlib==0.8.0
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tweepy==3.5.0
urllib3==1.22
